controller: Add a new callback to stop the event loop.
The event is mainly used for module detection.
The controller create a thread to run the loop and generate
the hotplug and hot unpulg events.
But an event loop may prevent the application to exit,
because it use methods that prevent thread to be cancelled.
Add a new callback that will be called before to cancel the
thread to put the event loop in a cancellable state.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>