Reset Greybus before to exit the application
Currently, when gbridge exit, it doesn't remove the SVC.
When gbridge will run again, it won't be able to initialize
the SVC because Greybus doesn't expect a new SVC to registered.
Greybus doesn't support SVC ho unplug, so to fix it, we have unload
greybus hd driver.
Use the GB_NL_C_HD_RESET to unload and then load the hd driver.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>