svc: Make development and debugging of SVC protocol easier
SVC operation are dispatched in SVC driver by switch case block.
It works fine but it is hard to figure out which operation handler
is complete, which one is just a stub method or which one is missing.
In addition, it's not really scalable. By example, to print then of
operation that failed, we will have to add and maintain another switch
case block that print operation name (instead of operation type).
Use new struct to register a callback and some other information about
operation.
Register all operation in an array and get the operation handler from it.
In the case of failure, the operation dispatcher will be able get some
additional information from the operation handler and give some details
about the error.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>