greybus: add a new macro operation_handler macro
Add the macro REQUEST_NO_HANDLER() to define an operation handler
without a callback. This will let us add new SVC operations without
to have to implement them.
Operation registered with REQUEST_NO_HANDLER() are not supposed
to be used by the kernel in the current state, but if the kernel
does, then gbridge will print an error.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>