controller: make get_interface() function public
In order to make the code more generic and runable on both AP and module
side, we are going to use the cport and interface id instead ofhd cport id.
But that means we need to be able to get the interface by using its id.
Make get_interface() function public.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>